requirements.txt
# Python 3.10+ Compatible Requirements
# Tested and working configuration for Mask R-CNN component segmentation

# Core numerical/scientific computing
numpy==1.23.5         # Critical version - see MIGRATION_NOTES.md
scipy>=1.10.0

# Deep learning (TensorFlow 2.11 is last version with standalone Keras)
tensorflow==2.11.0
keras==2.11.0
protobuf<4.0         # Required for TF 2.11

# Model persistence
h5py>=3.8.0

# Image processing
opencv-python>=4.7.0
Pillow>=9.5.0

# Visualization
matplotlib>=3.7.0

# Configuration
PyYAML>=6.0

# Testing (optional - for development only)
pytest>=7.0.0

# Note: scikit-image is NOT included due to compatibility issues
# Use OpenCV or Pillow for image I/O instead
# See MIGRATION_NOTES.md for details
